+++ Collectathon.Traits/Postpend.cs
@@ -10,7 +10,7 @@ namespace System.Traits {
 		/// <param name="count">The amount of elements in the collection; the amount currently in use.</param>
 		/// <param name="element">The element to postpend.</param>
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
-		public static void Postpend<TElement>([AllowNull] TElement[] collection, ref nint count, [AllowNull] TElement element) => Postpend(collection.AsSpan(), ref count, element);
+		public static void Postpend<TElement>([AllowNull] TElement[] collection, ref Int32 count, [AllowNull] TElement element) => Postpend(collection.AsSpan(), ref count, element);
 
 		/// <summary>
 		/// Postpends the elements onto this object, as a batch.
@@ -19,7 +19,7 @@ namespace System.Traits {
 		/// <param name="count">The amount of elements in the collection; the amount currently in use.</param>
 		/// <param name="element">The element to postpend.</param>
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
-		public static void Postpend<TElement>(Memory<TElement> collection, ref nint count, [AllowNull] TElement element) => Postpend(collection.Span, ref count, element);
+		public static void Postpend<TElement>(Memory<TElement> collection, ref Int32 count, [AllowNull] TElement element) => Postpend(collection.Span, ref count, element);
 
 		/// <summary>
 		/// Postpends the elements onto this object, as a batch.
@@ -28,7 +28,7 @@ namespace System.Traits {
 		/// <param name="count">The amount of elements in the collection; the amount currently in use.</param>
 		/// <param name="element">The element to postpend.</param>
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
-		public static void Postpend<TElement>(Span<TElement> collection, ref nint count, [AllowNull] TElement element) => collection[(Int32)count++] = element;
+		public static void Postpend<TElement>(Span<TElement> collection, ref Int32 count, [AllowNull] TElement element) => collection[(Int32)count++] = element;
 
 		/// <summary>
 		/// Postpends the elements onto this object, as a batch.
@@ -37,7 +37,7 @@ namespace System.Traits {
 		/// <param name="count">The amount of elements in the collection; the amount currently in use.</param>
 		/// <param name="elements">The elements to postpend.</param>
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
-		public static void Postpend<TElement>([AllowNull] TElement[] collection, ref nint count, ReadOnlySpan<TElement> elements) => Postpend(collection.AsSpan(), ref count, elements);
+		public static void Postpend<TElement>([AllowNull] TElement[] collection, ref Int32 count, ReadOnlySpan<TElement> elements) => Postpend(collection.AsSpan(), ref count, elements);
 
 		/// <summary>
 		/// Postpends the elements onto this object, as a batch.
@@ -46,7 +46,7 @@ namespace System.Traits {
 		/// <param name="count">The amount of elements in the collection; the amount currently in use.</param>
 		/// <param name="elements">The elements to postpend.</param>
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
-		public static void Postpend<TElement>(Memory<TElement> collection, ref nint count, ReadOnlySpan<TElement> elements) => Postpend(collection.Span, ref count, elements);
+		public static void Postpend<TElement>(Memory<TElement> collection, ref Int32 count, ReadOnlySpan<TElement> elements) => Postpend(collection.Span, ref count, elements);
 
 		/// <summary>
 		/// Postpends the elements onto this object, as a batch.
@@ -55,8 +55,8 @@ namespace System.Traits {
 		/// <param name="count">The amount of elements in the collection; the amount currently in use.</param>
 		/// <param name="elements">The elements to postpend.</param>
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
-		public static void Postpend<TElement>(Span<TElement> collection, ref nint count, ReadOnlySpan<TElement> elements) {
-			elements.CopyTo(collection.Slice((Int32)count));
+		public static void Postpend<TElement>(Span<TElement> collection, ref Int32 count, ReadOnlySpan<TElement> elements) {
+			elements.CopyTo(collection.Slice(count));
 			count += elements.Length;
 		}
 	}
